requirements.txt
flask>=2.0.0
pixivpy3>=3.7.0
pyyaml>=6.0
gppt>=1.0.0
